requirements.txt
django==4.2.5
Pillow==10.0.1
Scikit-learn>=1.3.1
numpy==1.26.0
requests